+++ Server/signal_server_demonstration/api/views.py
@@ -19,34 +19,63 @@ class MessageList(APIView):
     def get(self, request, **kwargs):
         user = self.request.user
 
+        # Check correct arguments provided
+        if not 'requestedDeviceRegistrationID' in kwargs:
+            return errors.incorrect_arguments
+
         # Check device exists and owned by user
         if not hasattr(user, "device"):
-            return Response("Device does not exist", status=status.HTTP_403_FORBIDDEN)
+            return errors.no_device
+
+        # Check device ID has not changed
+        if int(kwargs['requestedDeviceRegistrationID']) != user.device.registrationId:
+            return errors.device_changed
             
         messages = user.device.received_messages.all()
         serializer = MessageSerializer(messages, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     # User can post multiple messages. They will be defined as the sender
-    def post(self, request):
+    def post(self, request, **kwargs):
 
-        user = self.request.user
-        messageData = request.data
+        # Check correct arguments provided
+        if not 'requestedDeviceRegistrationID' in kwargs:
+            return errors.incorrect_arguments
+        if not (hasattr(request, "data") & isinstance(request.data, object)):
+            return errors.incorrect_arguments
+        if not (("recipient" in request.data) & isinstance(request.data["recipient"], str)):
+            return errors.incorrect_arguments
+        if not (("message" in request.data) & isinstance(request.data["message"], str)):
+            return errors.incorrect_arguments
+
+        ownUser = self.request.user
+        recipientUsername = request.data["recipient"]
+        messageData = request.data["message"]
+
+        # Check device exists and owned by user
+        if not hasattr(ownUser, "device"):
+            return errors.no_device
+
+        # Check device ID has not changed
+        if int(kwargs['requestedDeviceRegistrationID']) != ownUser.device.registrationId:
+            return errors.device_changed
+
+        # Check repipient user exists
+        try:
+            recipientUser = User.objects.get(username=recipientUsername)
+        except:
+            return errors.no_recipient
+
+        # Check recipient device exists
+        if not hasattr(recipientUser, "device"):
+            return errors.no_recipient_device
+        recipientDevice = recipientUser.device
+
         response = []
 
         for message in messageData:
 
-            # Get sender's device object
-            if not Device.objects.filter(address= message['senderAddress']).exists():
-                return Response("Sender device does not exist", status=status.HTTP_403_FORBIDDEN)
-            senderDevice = Device.objects.get(address= message['senderAddress'])
-
-            # Get recipient's device object
-            if not Device.objects.filter(address= message['recipientAddress']).exists():
-                return Response("Recipient device does not exist", status=status.HTTP_403_FORBIDDEN)
-            recipientDevice = Device.objects.get(address= message['recipientAddress'])
-
-            serializer = MessageSerializer(data=message, context={'senderDevice': senderDevice, 'recipientDevice': recipientDevice})
+            serializer = MessageSerializer(data=message, context={'senderDevice': ownUser.device, 'recipientDevice': recipientDevice})
             if not serializer.is_valid():
                 response.append(serializer.errors)
             else: 
@@ -56,24 +85,37 @@ class MessageList(APIView):
         return Response(response, status=status.HTTP_201_CREATED)
 
     # User can delete any message for which they are the recipient
-    def delete(self, request):
+    def delete(self, request, **kwargs):
+
+        # Check correct arguments provided
+        if not 'requestedDeviceRegistrationID' in kwargs:
+            return errors.incorrect_arguments
+        if not (hasattr(request, "data") & isinstance(request.data, list) & len(request.data) > 0):
+            return errors.incorrect_arguments
+
         user = self.request.user
         messageList = request.data
         response = []
 
+        # Check device exists and owned by user
+        if not hasattr(user, "device"):
+            return errors.no_device
+
+        # Check device ID has not changed
+        if int(kwargs['requestedDeviceRegistrationID']) != user.device.registrationId:
+            return errors.device_changed
+
         for messageId in messageList:
-            # Check message exists
-            if not Message.objects.filter(id=messageId).exists():
-                return errors.non_existant_message
-            else:
+            try:
                 message = Message.objects.get(id=messageId)
                 # Check user owns meaage
                 if not message.recipient.user == user:
-                    return errors.not_message_owner
-
-        for messageId in messageList:
-            message = Message.objects.get(id=messageId)
-            message.delete()
+                    response.append(errors.not_message_owner)
+                else:
+                    message.delete()
+                    response.append('success')
+            except:
+                response.append(errors.non_existant_message)
 
         return Response(response, status=status.HTTP_200_OK)
 
@@ -82,29 +124,29 @@ class DeviceView(APIView):
     # User can register details of a new device
     def post(self, request, **kwargs):
 
-        try:
+        # Check correct arguments provided
+        # Note - do not verify registrationID here as device should not exist
+        if not (hasattr(request, "data") & isinstance(request.data, object)):
+            return error.incorrect_arguments
 
-            user = self.request.user
-            registrationId = kwargs['deviceRegistrationId']
+        user = self.request.user
 
-            # Check device does not already exist
-            if hasattr(user, "device"):
-                return errors.device_exists
+        # Check device does not already exist
+        if hasattr(user, "device"):
+            return errors.device_exists
 
-            deviceData = request.data
-            serializer = DeviceSerializer(data=deviceData, context={'user': user})
+        deviceData = request.data
+        serializer = DeviceSerializer(data=deviceData, context={'user': user})
 
-            if not serializer.is_valid():
-                return errors.invalidData(serializer.errors)
-                
-            serializer.save()
-            return Response({"code": "device_created", "message": "Device successfully created"}, status=status.HTTP_201_CREATED)
-
-        except PermissionDenied:
-            return errors.reached_max_devices
+        if not serializer.is_valid():
+            return errors.invalidData(serializer.errors)
+            
+        serializer.save()
+        return Response({"code": "device_created", "message": "Device successfully created"}, status=status.HTTP_201_CREATED)
 
     # User can delete a device they own
     def delete(self, requested, **kwargs):
+        # Note - do not verify registrationID here as device should not exist
         user = self.request.user
         # Check device exists and owned by user
         if not hasattr(user, "device"):
@@ -118,18 +160,31 @@ class PreKeyBundleView(APIView):
     # User can optain a preKeyBundle from another user
     def get(self, request, **kwargs):
 
-        # Check user exists
-        if not User.objects.filter(username= kwargs['requestedUsername']).exists():
-                return Response("User does not exist", status=status.HTTP_403_FORBIDDEN)
+        ownUser = self.request.user
 
-        # Get user details object
-        user = User.objects.get(username=kwargs['requestedUsername'])
+        # Check correct arguments provided
+        if not (('recipientUsername' in kwargs) & ('ownDeviceRegistrationID' in kwargs)):
+            return errors.incorrect_arguments
 
-        # Check user has a registered device
-        if not hasattr(user, "device"):
+        # Check device exists and owned by user
+        if not hasattr(ownUser, "device"):
             return errors.no_device
 
-        device = user.device
+        # Check device ID has not changed
+        if int(kwargs['ownDeviceRegistrationID']) != ownUser.device.registrationId:
+            return errors.device_changed
+
+        # Check recipient user exists
+        recipientUser = {}
+        try:
+            recipientUser = User.objects.get(username=kwargs['recipientUsername'])
+        except:
+            return errors.no_recipient
+
+        # Check user has a registered device
+        if not hasattr(recipientUser, "device"):
+            return errors.no_recipient_device
+        device = recipientUser.device
 
         # Check prekey available for device before proceeding
         if device.prekey_set.count() == 0:
@@ -162,11 +217,21 @@ class UserPreKeys(APIView):
 
             user = self.request.user
 
-            # Check device exiss and owned by user
+            # Check correct arguments provided
+            if not 'requestedDeviceRegistrationID' in kwargs:
+                return errors.incorrect_arguments
+            if not (hasattr(request, "data") & isinstance(request.data, list) & len(request.data) > 0):
+                return errors.incorrect_arguments
+
+             # Check device exists and owned by user
             if not hasattr(user, "device"):
                 return errors.no_device
-                
-            newPreKeys = request.data['preKeys']
+
+            # Check device ID has not changed
+            if int(kwargs['requestedDeviceRegistrationID']) != user.device.registrationId:
+                return errors.device_changed
+            
+            newPreKeys = request.data
 
             for x in newPreKeys:
                 serializer = PreKeySerializer(data=x, context={'user': user, 'registrationId': user.device.registrationId})
@@ -188,18 +253,26 @@ class UserSignedPreKeys(APIView):
 
         user = self.request.user
 
+        # Check correct arguments provided
+        if not 'requestedDeviceRegistrationID' in kwargs:
+            return errors.incorrect_arguments
+        if not (hasattr(request, "data") & isinstance(request.data, object)):
+            return errors.incorrect_arguments
+
         # Check device exists and owned by user
         if not hasattr(user, "device"):
             return errors.no_device
+
+        # Check device ID has not changed
+        if int(kwargs['requestedDeviceRegistrationID']) != user.device.registrationId:
+            return errors.device_changed
             
-        device = user.device
-        newSignedPreKey = request.data['signedPreKey']
-        serializer = SignedPreKeySerializer(data=newSignedPreKey, context={'user': user, 'registrationId': device.registrationId})
+        serializer = SignedPreKeySerializer(data=request.data, context={'user': user, 'registrationId': user.device.registrationId})
 
         if not serializer.is_valid():
             return errors.invalidData(serializer.errors)
             
-        device.signedprekey.delete()
+        user.device.signedprekey.delete()
         serializer.save()
         return Response({"code": "signed_prekey_stored", "message": "Signed prekey successfully stored"}, status=status.HTTP_200_OK)
